Allow option alternatives after long options
This change makes the syntax support:
- long option followed by another long option (see highlighted/bundle)
- long options before short options (see grep, make, etc.)

It also creates false positives in explanation lines starting with a
long option (see the regression in delta and ripgrep). But it was
already detected (see syntax_test_bat#L25) so I won't add a new syntax
test for this.

I think I have a fix for these: limit definitions to indentations of <=8
spaces (in the line-indent context) so this regression is acceptable at
this point in time.