Limit length of section headings to 40 characters
To prevent the detection of plain explanation lines as section headings,
caused by the previous 2 commits.

40 chars are a lot but some section names are really long, e.g.: grep's
"Pattern selection and interpretation:" (37 chars).

I was stuck here for a long while cause I tried to write the
length-measuring lookahead like '^(?=.{,40}\n)', where {,40} is a
quantifier that means "at least 0 but not more than 40 times" according
to the Oniguruma regex reference [1].

But it seems that matching zero length inputs is not allowed in some
cases [2], so {1,40} is the right way to write it.

[1] https://raw.githubusercontent.com/kkos/oniguruma/v6.9.1/doc/RE
[2] https://github.com/trishume/syntect/issues/59#issuecomment-472718568