Make inline "usage:" scope more specific
My use case: in order to italicize it. It's still displayed like a normal
section heading for everyone else (no changes in tests/highlighted).

With this, section headings are handled pretty decently, and I plan to
shelve the todo's for this and focus on improving option detection and
other definitions (args, subcommands, etc.)