Dockerize syntect/syntest for syntax tests
The dockerfile:

- git clones trishume/syntect
- cargo builds the syntest example
- copies the executable to a much slimmer image (was full debian)
- runs syntest with syntax_tests in /tests/syntax and the syntax itself
  in /tests/syntax/under-test

What's not in the code? I built it from tests/ like this:

$ docker build -f syntest.dockerfile -t=syntest .

And it uses a volume like this (pwd is project root):

$ docker run --rm -v="$PWD/tests/syntax":/tests/syntax syntest

Also the syntax definition has to be copied to tests/syntax/under-test
in order to update it. Symlinks don't work within docker volumes. :/

I intend to write a Python wrapper around all of this too.