Limit line-indent to > 0, <= 8 chars
Two changes, a functional one and a little refactor.

1. Limit line-indent to <= 8 chars, which fixes some false positives
when explanation lines begin with an option (see modified syntax test).

2. Don't allow 0 chars line-indent's, because coherence. Instead push
the def-option context after line-begin, which is the context that
handles "0 chars line-indent's".

Only the python and the less help messages experience regressions. The
former is somewhat acceptable. We don't care about the less help message
cause it uses a weird encoding (groff or whatsitsname, the same as
manpages). I need to add a unit test for that.