Add regression tests for theme support
Until now we targeted the Monokai default theme only.

This means the syntax looks bad on quite a few themes because not all of
them target the scopes we use. Especially entity.name for option
definitions.

Now I'm going to tweak scopes (multi-scoping) and see how that affects
each theme.