Don't make tests depend on a ~/config/bat symlink
Removes the regression tests' dependency for a symlink
~/config/bat/syntaxes/cmd-help.sublime-syntax -> ./cmd-help.sublime-syntax

It requires placing the syntax in syntaxes/ to reproduce bat's config but
that shouldn't be a problem when PR-ing this as a bat submodule. See
sublimetext-asciidoc in [1].

As a nice side-effect, removes a bit of the indirection on the docker
volumes mounted in tests/syntax.py.

[1] https://github.com/sharkdp/bat/tree/master/assets/syntaxes/02_Extra