Try to get all themes to colorize arguments
This is a failed attempt to get option arguments colorized for all
themes by multi-scoping them with variable.other.

It's a failed attempt because it seems those themes that don't color
variable.parameter, it's not that they don't target the scope but that
they set it same as their default foreground color...

I'm committing this because I think the new architecture for declaring
scopes in an anchor and reusing that everywhere can be useful for future
attempts.

I didn't bother to change syntax tests though.