Multi-scope section headings as entity.name.section
On top of "markup.heading". "entity.name.section" with higher priority
because it's the standard scope for heading content.

For most changed themes, it adds color where there was none. Although
it's the same color as option names in most of them.

For ansi it's a regression: goes from one color to the same one as
options. For Solarized it's the opposite, so a net improvement.