Re-architecture (end): remove debug scopes
The highlight samples change because these debug scopes added color
escape characters to all text that was otherwise unscoped, but those
color escapes where foreground=white (like normal text).

The re-architecture is now complete. We're back to (almost) the same
result as when we started, with all the features intact and syntax tests
passing. But now the structural contexts do as they say and there will
be less head-scratchers thanks to that.

I'd like to add syntax tests to lock this important behavior: closing
the line properly, by popping the line context and resuming from main.
But it's hard to do so without adding a scope (meta.unimportant) and a
"rest-of-line" context for it, which I'd rather not do.

The right way to compare this is diff-ing its highlight samples with the
latest commit before beginning the re-arch:

$ GIT_PAGER="LESS=R less" git diff 6851257 -- tests/highlighted