Handle + sign as option lead
Not sure how to scope this. In the end I went with + being a keyword
and I plan to treat the bits that come after as an option argument.

Going with keyword.control rather than the more appropriate
keyword.operator cause keyword.operator is themed with foreground or
grey colors.