Embed man syntax if input looks like a man page
Damn, this one was tricky. I was unable to make it work until I removed
a `pop: 1` (or `clear_scopes: true`) that we otherwise need...

This works but it's sub-optimal, I'm unable to remove the txt.cmd-help
scope.

Other things that worked:

```yaml
match: man-regex
  push: scope:text.man
  # or
  push: Manpage.sublime-syntax
```

Didn't work:

```yaml
match: man-regex
  pop: 1
  # or
  clear_scopes: true
  # before embed/push
```