Comment alternative scope for other-def
I've tried support.option... instead of keyword.other, it may be a
better scope name considering most other-def matches will be
subcommands.

But it causes collision or similarity with the --option or heading
scopes, specially the former which is entity.name.option.

So leave it commented as an alternative, perhaps if/when I re-scope
option-def, but don't change it.