Add python to theme regression tests
Cause bat and vim don't have environment variables like PYTHONSTARTUP,
and I need to test how different scope names for this are colorized in
each theme.

The result for ansi is pretty weird, but it only happens with the python
help message and not with the already existing bat-short and vim. I may
want to investigate this later on.