Do away with non-branching maybe-option
No effective changes on sample messages.

This is the first step of a re-architecture motivated by two things:

As I extended maybe-argument, my first approach of matching a prefix
lookahead in maybe-X and committing (no backtrack) to the proper X
context, it was becoming cumbersome. Cause we can just match whatever in
the X context without having to keep track of valid prefixes in maybe-X.

Second, the uncaught command argument in syntax_test_emacs#L38 surfaced
an issue with how indent-0 commits to heading. It doesn't match
because that's a definition, and because it pops, it never tries the
actual argument-def context that should scope it.

This commit is a first step and a successful experiment on how enforcing
structure (the indent-X contexts) through includes works just as well as
setting contexts that won't backtrack. And the next steps should prove how
it's also infinitely more flexible.