Refactor heading regex
As expected, this fixes the issue in emacs's help message.

This preserves the headings with only 40 chars or less constraint by
placing the lookahead in the same regex that matches the heading. It
hadn't occurred to me that lookaheads don't only go at the tail of a
regex... :sweat_smile:

This tutorial is pretty useful:
https://www.regular-expressions.info/lookaround2.html