Don't abort on non-color codes
This fixes the case where a control sequence for bold text makes the
syntax stop operation. Now we only abort on actual color codes,
according to the escape sequences I found in [1].

Most of these color codes are going to go untested but we can live with
that.

It's not my ideal outcome, which would be scoping on top of the
pre-existing format (so color + bold/italic/etc), but that seems tricky
to make work. We would need to match and ignore escape sequences in
every little piece of regex. Unmaintainable.

[1] https://misc.flogisoft.com/bash/tip_colors_and_formatting