Handle "subcommand - description" from apt
There's a couple of unintended but happy side effects in sudo and jq,
where they have, e.g.: "sudo - execute a command as another user" and
the command name gets scoped.

And then there's a regression in this weirdly formatted enum definition
man has. But it's minimal so it's more or less acceptable.