Allow lowercase other-defs followed by 2 spaces
With caveats. The "or  " is worth an exception, and the negative
lookahead to prevent 2-spaces later in the same line is to avoid help
messages that have justified text, with words separated by 2 spaces.

Shitty thing is I cannot find where I saw one such command, but I've a
distinct memory that there was one. It's not on my notes either. :(

Either way, I'll err on the side of caution here, with the lookahead.