Handle 'argument' definitions
Taken from tee, where they're actually an enum for the values of an
option argument.

But I'm not too clear what they should be in general. For the time being
I set it to other_variable_def_scope, but I think I should consolidate
that one (or its color) with option and command arguments.