Drop unnecessary quotes around string scalars
Most of them.

I'm keeping the ones in (regex) variables because most variables begin
with '[' or '{' and those require surrounding quotes. The ones where
it's not necessary, I keep the quotes for consistency.

Similar reasoning for connector contexts (e.g.: option-def-connectors)
where a lot of connectors begin with whitespace or symbols that make
quotes necessary (e.g.: ',' or ':').

I'm actually not too sure about this change but I want to try. I can
always revert if I'm not comfortable with this.