Make tests/syntax.py more robust
The man syntax is required for syntect syntax tests because there's a
test (git-commit) where control is handed over to that syntax.

I'd planned to do this ages ago. I'm doing it now, in case someone feels
like tinkering with the project. Or in case I get down to add CI to
foster contribution.