Add bat-server.dockerfile
Pending a Compose file defining the required volumes.

This is a first step towards dockerizing bat, which will prevent my
messing with the local instance's cache.

Pain point: when I'm in the middle of making big changes to
custom16.tmTheme, without re-building the bat cache yet. Then the
"restore user configuration" step in regression scripts updates the
cache with those wip changes.

As a rule of thumb, messing with the user's (contributor's?) local
setup is a bad thing.

Credit for the CURL options:
https://veducate.co.uk/download-releases-from-github-using-curl-and-wget/