syntax.py: build syntest without build context
The previous build context (tests/docker/) was an arbitrary one set so
its size was minimal and builds ran faster.

The syntest image doesn't really require a build context. But it's a
required argument in docker build so there was no way around it.

But there's actually a better pattern for that: using stdin as the build
context, and piping the Dockerfile there.

Discovered in [1], with an improved version in the official docs [2].

[1] https://alecthegeek.github.io/docker/2019/06/06/Docker-Build-With-No-Build-       ﬌ Context.html
[2] https://docs.docker.com/engine/reference/commandline/build/#text-files