Drop docker compose, use layer cache instead
This drops the bat-server setup with a docker-compose in favor of a not
so over-engineered dockerfile.

This leverages the docker layer cache on each image re-build, so there's
no need to run `bat cache --build` if the syntax didn't change.

The point of bat-server was being able to use it for both highlight_ and
theme_regression.sh. But we can do that with a plain old Docker image if
we overwrite the entrypoint (run_highlight_ vs run_theme_regression.sh).

This improves DX for two reasons:

1. no need to `docker compose down` at the end of a coding session. (I
always forget.)
2. doesn't `bat cache --build` when the syntax file didn't change, which
makes that case speedup 2-3x.