Provide more context to *_regression.sh output
The sha256:xxxxxxxxxx output was kinda decontextualized, this provides
some context.

Tempted to just output to /dev/null but the SHA may be a good clue when
you're not doing what you want. E.g.: testing the syntax but you forgot
to save the file, then you get the same SHA as the last execution.