Add development scripts
These are super helpful to my work here, hopefully they'll also be to
potential contributors.

In particular, the aliases for diff-ing tests are pretty handy because I
have git diff tuned to show color with git-delta and these unset that.

And make_syntax_test is a hell of an over-engineered solution to a task
that I hated when I did manually. Because I always forgot to add the
SYNTAX TEST header and syntest would give me weird errors.