Add .gitattributes so GitHub adds YAML to langs
I discovered I could do this in the bat repo [1]. Neat!

This gives a much more faithful breakdown of what's in this project.
YAML makes up 63% of the code.

Further documentation on how to use .gitattributes to override
github-linguist in [2].

But there's a gotcha! github-linguist doesn't pick this up (or any other
change really), until it's committed. For trial & error of changes, you
need to commit amend.

[1] https://github.com/sharkdp/bat/blob/master/tests/.gitattributes
[2] https://github.com/github/linguist/blob/master/docs/overrides.md