Don't scope section headings in synthetic test
The purpose of that test is surfacing changes in how each theme
colorizes each token.

I'm working in an overhaul of scopes (and so color schemes), starting
from the heading scopes, and these section headings keep showing in the
diff when they shouldn't. I want concise diffs and these lines show up
in disjoint regions, making the diffs sparse. Shitty.

So I'm changing them to "## SECTION HEADING" because that's not
currently scoped as heading. If it ever does, we ought to change the
prefix characters.

I also hope this change is going to help the diff between main and that
"overhaul color schemes" branch, make it much more to the point.