Change arguments' scope to constant.other
The main improvement this gives is it fixes no-color arguments in base16
and arguments=headings in ansi, which are two themes I definitely wanted
to support. Because they use the terminal's color palette, instead of
arbitrary colors.

It also brings back some cool colors I lost on changing the heading
scope: blue on Coldark, orange on Solarized. It even improves support
for gruvbox (but still bad).

It's kind of weird to see Monokai arguments being purple instead of
brown-orange, but it makes sense in the context of the larger body of
themes. I will reuse that color (variable.parameter) on files' scopes
probably.

Things I tried:

- comment: bad, subtle hues on most themes
- support.type: bad
- constant: not targeted by many themes
- constant.numeric: may be an option
- constant.other: an excellent option
- constant.language: worse than .other
- constant.character: same
- storage.type: not particularly good
- storage.modifier: bad, too similar to keyword (for subcommands)
- support.constant: not particularly good
- support.function: bad, too similar to entity.name.function (for options)
- support.class: not particularly good
- support.type: bad
- support.module: bad
- keyword.control: bad, mostly like keyword.other (for subcommands)
- keyword.operator: same
- keyword.declaration: same
- variable.function: bad, often similar to entity.name.function
- variable.language: bad, no support on Monokai
- variable: bad, not matched by many themes
- entity.name.tag: often similar to keyword, not good enough.

Recap of scopes that I may use elsewhere:

- variable.parameter
- entity.name.tag
- constant. variants
- keyword. variants

I basically combed through the common scope names here:
https://www.sublimetext.com/docs/scope_naming.html