Go back to running syntax.py in CI (cache 8)
The previous commit worked! The syntest image is properly generated and
registered with `docker images`!

But the job still fails because we need the part of syntax.py where it
downloads the man syntax. Some syntax tests depend on that syntax being
available in syntaxes/.

Now, as the syntest image is already built when we hit syntax.py, it
skips rebuilding the image. Which is par for the course: caching, based
on docker/build-push-action.

ETA: indeed, this version passes :tada: :rocket::rocket: