Swap argument & subcommand scopes
i.e.: `constant.other` (A) for subcommands and `keyword.other` (B) for
arguments & option arguments.

Because, on most themes [3], the colors for A coordinate better with
the colors for options: `entity.name.function` (C). As compared to the
colors for B.

Particularly, one or both of the following hold:

1. A's hue has more contrast to C than B, which helps make apart options
and option arguments.
2. B is significantly more vivid* than A, sometimes even more than C,
which puts the focus on option arguments and detracts from options. But
options are more important than their arguments.

* Vivid as in brighter / more saturated.

For (2) I also fancy that subcommands are displayed in a more vivid
color than arguments and option arguments.

[3] A recap of how the change works on most themes:

- Coldark *: (1) applies
- Dracula: neutral?
- GitHub Light: 2
- Monokai *: 2
- Monokai Extended Light: the opposite is true, worse off
- OneHalfDark: 2
- OneHalfLight: 2
- TwoDark: 2
- Solarized *: 2
- Snazzy: 2
- Visual Studio Dark+: 1
. ansi: 2 but may lose out a bit on 1 (yellow is usually brighter / more
    saturated than magenta; but the later is closer to blue)
. base16: idem but swap yellow for red.
- gruvbox: neutral?
- zenburn: loses out on 1, but makes subcommands bold, which is better
    than option arguments