Make themes' quick reference into its own test
While working on overhaul-color-theme-support, I've realized the
quick reference section in the synthetic test for themes should be its
own test.

This should massage the diffs, lets me color test a change much quicker,
while quickly iterating on themes, etc.

I've decided to keep the italics version (another point of friction) in
the synthetic test, because italics isn't nearly as important in checks
as color is.

This also revealed a little problem in my dockerized test setup. The
highlighted files are generated with root permissions! It's not a big
problem with git, but when you want to delete files and it asks for
permissions it's not cool.