Add ad-hoc scope to make headings golden in Monokai
Context: Monokai Extended is the default theme for bat. It's also the
theme we use for highlight regression tests.

Earlier in this overhaul of scopes, I changed the scope for headings in
order to better support most themes, particularly themes based on the 16
customizable terminal colors (ansi & base-16), which previously used the
same color for headings & options.

Before that change, Monokai used the same orange color for both headings
and arguments (which is bad). After it, headings became a pale yellow.

Later on, I also changed the scope for arguments, because it caused
color repetitions in other themes I also care about.

So in the end, for the default Monokai theme, we went from an overuse of
the orange color to an absence of it. Which looks weird to me, I'm used
to the orange.

Furthermore, and perhaps more relevant, I'm not much a fan of the pale
yellow color that headings took. I don't think it helps highlight them
over other elements (options in particular), which is bad.

So I looked for orange color assignations in the theme [1], and found
the previous orange is a bit hard to target without messing up other
themes. But there's this alternative hue of orange that's used on a few
scopes, including this very specific one that other themes don't target:
`meta.constructor.argument.css`. I don't even know what that is, Google
didn't help.

Coincidentally, this is a slightly different hue of orange, closer to
gold, with more brightness. Which is great for highlighting headings
over other elements. At the same time, it's close enough to the previous
orange hue that users (or me) won't feel the difference unless they're
looking for it.

It's also closer to my preferred shade of yellow (cf. the change on
docs/Principles.md).

ETA: another nice side effect I noticed testing this against highlighted
regression tests: this orange-golden color is closer to what Manpage
headings get with that syntax (it adds a scope that's painted with the
previous orange). The pale yellow looks weird next to that.

The only other built-in theme where this tweak causes a diff is Dark
Neon, where headings become a darker green (worse). But it's an
acceptable regression because Dark Neon was badly supported already
(options use the default foreground color).

If this hack causes problems with other built-in themes down the line,
I'm open to reverting it. For now, it's a net improvement in my book.

[1] https://github.com/jonschlinkert/sublime-monokai-extended/blob/master/Monokai%20Extended.tmTheme