Fix bug in OPTION_ARGUMENT_SCOPES
`option.name` is not a thing, `entity.name` is.

Fixing this, a previous bug rears its head. But I already had a solution
in a stash so :shrugs:

The bug: de-scope by `clear_scopes: 1` doesn't work if the outer context
assigns a composite scope (like `entity.name.function variable.parameter`).

To properly clear that (and go back to default foreground color) you
need to clear all the scopes except the base one (`text.cmd-help`)

Which means that to properly do this in arguments (1 scope as opposed
to 2 from option arguments), we should only pop 1 scope, not 2. This is
not a problem in most themes, only in Gruvbox (which is broken anyways),
so not a priority to properly handle this.

The non-wrong way to do this is having a de-scope context for contexts
that add 1 scope, 2 scopes, etc. As long as these context have to
de-scope at some point.

The correct way to do this is devising an entirely different algorithm,
but refactoring this hack is not a priority