Add hyper-specific scope to FILE_DEF for Monokai
This changes the color for this element in Monokai, from orange (not the
same as headings golden, but pretty close to) to blue.

This affects all of the Monokai variants except Monokai Extended Bright.

There's also some weird bug in ansi with how the color is not properly
cleared after the token and it propagates to the "bla bla bla"
explanation.

I don't know if the error is on bat/syntect's side or ours.