Update expectations for syntest caching in CI
It alarmed me that the syntest CI step the other day took >3m, which was
supposed to be fixed with the docker/build-push-action that caches the
result of the docker build in GitHub's cache.

I've looked a bit into it and the docs [1] do state that GH removes cache
entries that haven't been accessed in over 7 days. I hadn't worked on
cmd-help since a month ago. Bummer.

I could make a dummy PR, to be pushed to every week, to touch the cache.
But that solution is worse than the problem, and wasteful as heck.

So we'll have to live with cold starts of "CI / Syntax tests", if we
sleep on the project for more than a week.

[1] https://docs.github.com/en/actions/using-workflows/caching-dependencies-to-speed-up-workflows#usage-limits-and-eviction-policy