syntest: Revert to building with release profile
It fails to compile with the dev profile. Changed in 2517499

Also, bump the base image for the build from rust:1.59 to 1.65.

Because the build failed from a couple dependencies not working with
that version of rustc:

error: package `csv v1.2.2` cannot be built because it requires rustc 1.60 or newer, while the currently active rustc version is 1.59.0
error: package `time v0.3.21` cannot be built because it requires rustc 1.65.0 or newer, while the currently active rustc version is 1.62.1

It's been working all this time in my local, maybe, because it was using
cached versions of the syntest build.