Handle ""pair HOST[:PORT] ..." from adb
Only the "pair" bit.

This is all so overfitted, to suit adb's help message only, that it's
making me feel icky.

But, on the other hand, it's also a good census of all kinds of patterns
that we don't currently have a proper way to handle.

And if it doesn't add too many unnecessary scopes for other help
messages, I don't care much, I guess?