Handle unmatched option alt patterns in wget
The problem was with how wget has ",  " (two spaces) as a connector
between alternative option names.

Not to happy to add this uncommon pattern as the first one in
option-def-alt but cannot avoid this if we don't want to match these as
", " (single space) and then fail the matches in the option-term
context.

Fixes #22.