Separate {{long_space}} and {{single_2_space}}
I hoped to catch the tricky chown case with this, but I think I know why
it's failing: we do not `set: then-pop" in keyword-seq-def matches !!

As a side effect, we scoped "operand=VALUE" operands in dd, which is
nice.

As another side effect, we scoped "or:" usage lines in vim, which is bad
but easy to fix (next commit).