Fix regex for lookahead_single_2_space
Because it was matching the 2-space with the first lookahead, without
consuming the characters it matched. Then it matched the second
(negative) lookahead with the same 2-space. But because it's negative,
the whole match failed.

We just have to circumscribe the 2nd lookahead (negative) within the
first one.

This catches the 2 stray corner cases that I wanted to fix before
merging this branch: one line each in chown and npm.

But it also adds a hell of a regression in that it now matches plain
text from a lot of help messages. But this can be fixed easily enough,
cf. next commit.