Change some script code for clarity
Rename the volume variables for scripts using the bat-test image so it's
more clear what the container does.

Also, add an entrypoint for the docker run call on highlight_regression.
It's redundant because it's exactly the same as declared in the bat-test
image. But explicit > implicit.

It's still worth it that the Docker image does highlight regression by
default, if we want to use/test it outside of any script.