Fix bat cache permissions in bat-test container
This fixes the parent commit.

I hadn't tested properly and it turns out that the container was running
without using the re-built assets cache that includes our version of
cmd-help.sublime-syntax, the one we should be testing.

I had tested that, after deleting the files in tests/theme/, bat was
able to re-generate those same files. And it did (git showed no diff)
but by using the cmd-help.sublime-syntax bundled with bat itself,
instead of the one under test.

This is because on the `bat cache --build` step, executed as root within
the image's intermediate steps, bat stores the compiled custom assets in
/root/.cache/bat/.

But then we run the container with the host sytem's user (rootless), and
bat has 1) no knowledge that the custom assets are in that path, and 2)
no permissions to read it because it's in root space.

I temporarily set the cache build target to the location where bat
expects it when running with the rootless user. As this user has no home
directory (the user is listed as "I have no name!", is it unknown?), the
expected location is /.cache/bat.

There's an alternative solution by setting the BAT_CACHE_PATH
environment variable in the Dockerfile, so root stores custom assets in
a directory accessible by a rootless user, and also so bat knows to find
them there when run by an user different than the one that rebuilt the
bat cache.

But I want to commit this no-frills solution first because the bat env
vars I want to use for the expressive one, they're not documented, they
may be deprecated down the line.

I'm considering that perhaps I should not flush (> /dev/null) the output
of commands like `docker build` or `bat cache --build`, but rather
redirect it to some logfile. It mutes it alright but at the same time it
keeps it available if I need to debug the image.