Use bat environment variables within bat-test image
For the cache directory (BAT_CACHE_PATH) and for the config directory
(BAT_CONFIG_DIR).

This is a more expressive & self-documenting solution than the approach
in the parent commit.

It should also make the container agnostic of whether it's run as root
or as a rootless user, cause they both save into and look at the same
config/cache locations.

It also makes all these files easily available under the bat root
directory, which makes it easy to debug the setup when things go awry.

It incurs the risk that bat deprecates these env variables, as explained
in the parent commit's message. But whatever, we can always rollback to
that.