Check compliance with bat regression tests on CI
Bat tracks syntax regressions with a test file for each syntax. In our
case, that sample file is the short help for bat v0.22.1.

We already tracked that sample in our own highlight regression tests,
but now we make this check more explicit through an extra job in our
GitHub Actions CI.

This new job doesn't fail the whole CI workflow, but emits a warning
when our highlight for this sample differs from the one in bat. Then it
links our docs with the explanation on why/how to update the regression
test in the bat project.