Readme: rewrite Configuration section
Now that we have proper documents for this in the project wiki [1,2], I
prefer the Readme only contains a teaser with links to those.

Single source of truth and all that jazz.

I also like the Wiki much better for docs, because it makes editing more
thoughtless: Whenever I see a potential improvement, small as it may be,
I make a quick edit and move along.

As opposed to the current MO, where I'll consider whether I should
really make a commit for that small an edit, and often wait until I have
a bunch of edits ready and group them in a single commit or branch.

[1] https://github.com/victor-gp/cmd-help-sublime-syntax/wiki/Configuration-tips
[2] https://github.com/victor-gp/cmd-help-sublime-syntax/wiki/Bat-theme-support