Change demo_themes dest directory
No point in having so much nesting methinks.

Also, I gotta create the directory in advance, cause when Docker creates
it for me (as a volume) it does so as root. Then with the inner
container running as an unprivileged user, the container errors when
trying to write to that root-owned dir.