Add try_input.sh script
This makes it easy to run any help message by the local version of
cmd-help.

*Local: the latest version still not in a bat release, a new change
I'm testing on a branch, etc. Anything that's on the git working
directory really.

Previously, we'd have to copy the help message into a file in tests/src/
and run tests/highlight_regression.sh to "quickly" test something new.