Add fish regression test + syntax test
With inconsistencies reported in #26.

There are a bunch of false positives for 'keyword scopes' that don't show
in the issue's screenshots (an earlier bat/cmd-help version, I'd say).
These are probably derived from my work on #23. For me, they're the most
worrying part.

The ask about scoping examples like [^1] may be feasible after we do
usage lines (#24). The key part is the `>` character.

The stray 'option scope' false positive is puzzling on first look, can't
say why the syntax isn't detecting that it's plain text. I should fix it
if it's easy to do so.

Aside from these false positives here and there, the bathelp
highlighting for the message looks pretty well.

[^1]   > fish --debug=iothread