Remove vendored Manpage.sublime-syntax
This broke the bat project's Dependabot workflow to update cmd-help to
the latest commit. And for good reason.

Basically, the Manpage syntax we vendored here would conflict with the
one they host there. I don't even know which one would take precedence.
Thankfully it was caught by their CI.

The change was made on 39f2c5e.

The alternative solution here doesn't quite solve the original problem
but I guess it's better than downloading the file from the python script.

Docker cache should help a bit on the CI side, and we won't need a copy
of the syntax in our local environment.